pyproject.toml
[tool.poetry]
name = "konviktion-server"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
django-taggit = "^4.0.0"
channels = { extras = ["daphne"], version = "^4.0.0" }
omegaconf = "^2.3.0"
channels-redis = "^4.1.0"
django-guardian = "^2.4.0"
strawberry-graphql-django = ">=0.28"
pytest-asyncio = "^0.21.1"
django-choices-field = "^2.2.2"
pydantic = "<2"
django-simple-history = "^3.3.0"
psycopg = ">=3.1.8"
django-probes = "^1.7.0"
django-cors-headers = "^4.2.0"
authentikate = ">=0.1.11"
kante = ">=0.1.7"
koherent = ">=0.1.2"
django-redis = "^5.4.0"
notion-client = "^2.2.1"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "strawberry.ext.mypy_plugin"]
ignore_missing_imports = true
strict = true


[tool.django-stubs]
django_settings_module = "konviktion_server.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "konviktion_server.settings_test"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.280"
black = "^23.7.0"
pytest-django = "^4.5.2"
mypy = "^1.4.1"
django-stubs = "^4.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "*/migrations/**",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"
